refactor: update folder name to data_cache
