+++ lib/src/core/config/cache_config.dart
@@ -1,8 +1,9 @@
-import '../../modules/cache/domain/constants/cache_constants.dart';
-import '../../modules/cache/domain/enums/invalidation_type.dart';
-import '../../modules/cache/domain/value_objects/cache_size_options.dart';
-import '../../modules/cache/domain/value_objects/invalidation_methods/implementations/ttl_invalidation_method.dart';
-import '../../modules/cache/domain/value_objects/invalidation_methods/invalidation_method.dart';
+import '../../modules/data_cache/domain/constants/cache_constants.dart';
+import '../../modules/data_cache/domain/enums/invalidation_type.dart';
+
+import '../../modules/data_cache/domain/value_objects/cache_size_options.dart';
+import '../../modules/data_cache/domain/value_objects/invalidation_methods/implementations/ttl_invalidation_method.dart';
+import '../../modules/data_cache/domain/value_objects/invalidation_methods/invalidation_method.dart';
 
 class CacheConfig {
   final CacheSizeOptions sizeOptions;
@@ -17,14 +18,14 @@ class CacheConfig {
 
   factory CacheConfig.defaultConfig() => CacheConfig(invalidationMethod: const TTLInvalidationMethod());
 
+  InvalidationType get invalidationType => invalidationMethod.invalidationType;
+
   bool get isDefaultConfig {
     final isDefaultInvalidation = invalidationType == CacheConstants.defaultInvalidationType;
     final isDefaultCacheSizeOptions = sizeOptions == CacheSizeOptions.createDefault();
 
     return isDefaultInvalidation && isDefaultCacheSizeOptions;
   }
-
-  InvalidationType get invalidationType => invalidationMethod.invalidationType;
 
   Duration get expireMaxDuration {
     if (invalidationType == InvalidationType.ttl) {
