+++ lib/src/core/core.dart
@@ -1,4 +1,4 @@
-export '../modules/cache/domain/constants/cache_constants.dart';
+export '../modules/data_cache/domain/constants/cache_constants.dart';
 export 'config/cache_config.dart';
 export 'exceptions/auto_cache_manager_exception.dart';
 export 'services/directory_provider/directory_provider_service.dart';
