+++ lib/src/modules/data_cache/domain/dtos/clear_cache_dto.dart
