+++ lib/src/modules/data_cache/domain/dtos/get_cache_dto.dart
