+++ lib/src/modules/data_cache/domain/dtos/save_cache_dto.dart
