+++ lib/src/modules/data_cache/domain/entities/cache_entity.dart
