+++ lib/src/modules/data_cache/domain/enums/invalidation_type.dart
