+++ lib/src/modules/data_cache/domain/exceptions/cache_size_analyzer_exceptions.dart
