+++ lib/src/modules/data_cache/domain/exceptions/invalidation_methods_exceptions.dart
