+++ lib/src/modules/data_cache/domain/repositories/i_cache_repository.dart
