+++ lib/src/modules/data_cache/domain/types/cache_types.dart
