+++ lib/src/modules/data_cache/domain/usecases/clear_cache_usecase.dart
