+++ lib/src/modules/data_cache/domain/value_objects/cache_cryptography_options.dart
