+++ lib/src/modules/data_cache/domain/value_objects/cache_size_options.dart
