+++ lib/src/modules/data_cache/domain/value_objects/invalidation_methods/implementations/ttl_invalidation_method.dart
