+++ lib/src/modules/data_cache/domain/value_objects/invalidation_methods/invalidation_method.dart
