+++ lib/src/modules/data_cache/external/adapters/cache_adapter.dart
