+++ lib/src/modules/data_cache/external/adapters/enums/invalidation_type_adapter.dart
