+++ lib/src/modules/data_cache/presenter/controllers/base_cache_manager_controller.dart
@@ -1,4 +1,4 @@
-import 'package:auto_cache_manager/src/modules/cache/domain/dtos/get_cache_dto.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/dtos/get_cache_dto.dart';
 import 'package:flutter/material.dart';
 
 import '../../../../auto_cache_manager_initializer.dart';
