+++ test/auto_cache_manager_test.dart
@@ -1,5 +1,5 @@
 import 'package:auto_cache_manager/auto_cache_manager.dart';
-import 'package:auto_cache_manager/src/modules/cache/presenter/controllers/base_cache_manager_controller.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/presenter/controllers/base_cache_manager_controller.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:shared_preferences/shared_preferences.dart';
