+++ test/src/core/config/cache_config_test.dart
@@ -1,6 +1,6 @@
 import 'package:auto_cache_manager/src/core/config/cache_config.dart';
-import 'package:auto_cache_manager/src/modules/cache/domain/enums/invalidation_type.dart';
-import 'package:auto_cache_manager/src/modules/cache/domain/value_objects/cache_size_options.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/enums/invalidation_type.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/value_objects/cache_size_options.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
