+++ test/src/modules/cache/domain/constants/cache_size_constants_test.dart
@@ -1,4 +1,4 @@
-import 'package:auto_cache_manager/src/modules/cache/domain/constants/cache_size_constants.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/constants/cache_size_constants.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
