+++ test/src/modules/cache/domain/usecases/clear_cache_usecase_test.dart
@@ -1,7 +1,7 @@
 import 'package:auto_cache_manager/src/core/core.dart';
-import 'package:auto_cache_manager/src/modules/cache/domain/dtos/clear_cache_dto.dart';
-import 'package:auto_cache_manager/src/modules/cache/domain/repositories/i_cache_repository.dart';
-import 'package:auto_cache_manager/src/modules/cache/domain/usecases/clear_cache_usecase.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/dtos/clear_cache_dto.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/repositories/i_cache_repository.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/usecases/clear_cache_usecase.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
