+++ test/src/modules/cache/domain/usecases/save_cache_usecase_test.dart
@@ -1,13 +1,13 @@
 import 'package:auto_cache_manager/auto_cache_manager.dart';
 import 'package:auto_cache_manager/src/core/core.dart';
-import 'package:auto_cache_manager/src/modules/cache/domain/dtos/get_cache_dto.dart';
-import 'package:auto_cache_manager/src/modules/cache/domain/dtos/save_cache_dto.dart';
-import 'package:auto_cache_manager/src/modules/cache/domain/entities/cache_entity.dart';
-import 'package:auto_cache_manager/src/modules/cache/domain/enums/invalidation_type.dart';
-import 'package:auto_cache_manager/src/modules/cache/domain/enums/storage_type.dart';
-import 'package:auto_cache_manager/src/modules/cache/domain/repositories/i_cache_repository.dart';
-import 'package:auto_cache_manager/src/modules/cache/domain/services/invalidation/invalidation_cache_context.dart';
-import 'package:auto_cache_manager/src/modules/cache/domain/usecases/save_cache_usecase.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/dtos/get_cache_dto.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/dtos/save_cache_dto.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/entities/cache_entity.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/enums/invalidation_type.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/enums/storage_type.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/repositories/i_cache_repository.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/services/invalidation/invalidation_cache_context.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/usecases/save_cache_usecase.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
