+++ test/src/modules/cache/external/adapters/cache_adapter_test.dart
@@ -1,11 +1,11 @@
 // ignore_for_file: require_trailing_commas
 
 import 'package:auto_cache_manager/src/core/extensions/map_extensions.dart';
-import 'package:auto_cache_manager/src/modules/cache/domain/entities/cache_entity.dart';
-import 'package:auto_cache_manager/src/modules/cache/domain/enums/invalidation_type.dart';
-import 'package:auto_cache_manager/src/modules/cache/domain/enums/storage_type.dart';
-import 'package:auto_cache_manager/src/modules/cache/external/adapters/cache_adapter.dart';
-import 'package:auto_cache_manager/src/modules/cache/external/adapters/enums/invalidation_type_adapter.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/entities/cache_entity.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/enums/invalidation_type.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/enums/storage_type.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/external/adapters/cache_adapter.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/external/adapters/enums/invalidation_type_adapter.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
