+++ test/src/modules/cache/external/adapters/enums/invalidation_type_adapter_test.dart
@@ -1,5 +1,5 @@
-import 'package:auto_cache_manager/src/modules/cache/domain/enums/invalidation_type.dart';
-import 'package:auto_cache_manager/src/modules/cache/external/adapters/enums/invalidation_type_adapter.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/enums/invalidation_type.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/external/adapters/enums/invalidation_type_adapter.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
