+++ test/src/modules/cache/presenter/base_cache_manager_controller_test.dart
@@ -1,14 +1,14 @@
 import 'package:auto_cache_manager/auto_cache_manager.dart';
 import 'package:auto_cache_manager/src/core/core.dart';
 import 'package:auto_cache_manager/src/core/exceptions/initializer_exceptions.dart';
-import 'package:auto_cache_manager/src/modules/cache/domain/dtos/get_cache_dto.dart';
-import 'package:auto_cache_manager/src/modules/cache/domain/dtos/save_cache_dto.dart';
-import 'package:auto_cache_manager/src/modules/cache/domain/entities/cache_entity.dart';
-import 'package:auto_cache_manager/src/modules/cache/domain/enums/storage_type.dart';
-import 'package:auto_cache_manager/src/modules/cache/domain/usecases/clear_cache_usecase.dart';
-import 'package:auto_cache_manager/src/modules/cache/domain/usecases/get_cache_usecase.dart';
-import 'package:auto_cache_manager/src/modules/cache/domain/usecases/save_cache_usecase.dart';
-import 'package:auto_cache_manager/src/modules/cache/presenter/controllers/base_cache_manager_controller.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/dtos/get_cache_dto.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/dtos/save_cache_dto.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/entities/cache_entity.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/enums/storage_type.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/usecases/clear_cache_usecase.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/usecases/get_cache_usecase.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/domain/usecases/save_cache_usecase.dart';
+import 'package:auto_cache_manager/src/modules/data_cache/presenter/controllers/base_cache_manager_controller.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
